refactor: update content components
