+++ src/components/content/blog-card.tsx
