+++ src/components/content/mdx-components.tsx
@@ -1,7 +1,6 @@
 import Image from 'next/image';
 
 import { LinkPreview } from '@/components/content/link-preview';
-import { MdxCard } from '@/components/mdx-card';
 import { Callout } from '@/components/shared/callout';
 import { cn } from '@/lib/utils';
 import { MDXComponents } from 'mdx/types';
@@ -193,6 +192,5 @@ export const components: MDXComponents = {
   ),
   Image,
   Callout,
-  Card: MdxCard,
   LinkPreview,
 } as MDXComponents;
